Reorganize comment view files.
h/t @cmc333333
